+++ packages/react-components/src/components/shared/floating-source.tsx
@@ -2,20 +2,20 @@
 import { UseInteractionsReturn } from "@floating-ui/react";
 import { cloneElement, isValidElement, ReactElement, Ref } from "react";
 
-type TooltipSourceProps = {
+type FloatingSourceProps = {
   children: ReactElement<{ ref: Ref<any> }>;
   ref: Ref<any>;
   getSourceProps: UseInteractionsReturn["getReferenceProps"];
 };
 
-export const TooltipSource: React.FC<TooltipSourceProps> = ({
+export const FloatingSource: React.FC<FloatingSourceProps> = ({
   children,
   getSourceProps,
   ref,
 }) => {
   if (!isValidElement(children)) {
     throw new Error(
-      "[@yamori-design/react-components][Tooltip] The child must be a valid React element."
+      "[@yamori-design/react-components] The child must be a valid React element."
     );
   }